requirements.txt
fpdf
fpdf2
numpy
pandas
python-dotenv
Requests
seaborn
streamlit
streamlit
ydata_profiling
ydata_profiling
matplotlib
